Introduce Caching service
